Match initrd to vmlinuz by version rather than assume they're in sync and "latest" means the same for both files. Execute depmod "$Kver" and update-initramfs -k "$Kver" -u explicitly.

+++ src/UbuntuDiskless/Diskless_Setup.sh
@@ -52,23 +52,32 @@ function setup_tftp_boot
 		#ln -s ${Moon_RootLocation}/boot/initrd.img /tftpboot/${Moon_DeviceID}/initrd.img
       		
 		## Find kernel and initrd
-        	Kernel=
-        	Initrd=
-        	for File in /usr/pluto/diskless/${Moon_DeviceID}/boot/vmlinuz-*; do
-                	Kernel="$File"
-        	done
-        	for File in /usr/pluto/diskless/${Moon_DeviceID}/boot/initrd.img-*; do
-                	Initrd="$File"
-        	done
+		Kernel=
+		Initrd=
+		for File in /usr/pluto/diskless/"$DeviceID"/boot/vmlinuz-*; do
+			Kernel="$File"
+		done
+		if [[ -z "$Kernel" ]]; then
+			echo "WARNING: Missing kernel or initrd file. Cannot set for PXE boot."
+			continue
+		fi
+		Kver=$(basename "$Kernel")
+		Kver="${Kver#vmlinuz-}"
+		Initrd=/usr/pluto/diskless/"$DeviceID"/boot/initrd.img-"$Kver"
 
-        	if [[ -z "$Kernel" || -z "$Initrd" ]]; then
-                	echo "WARNING: Missing kernel or initrd file. Cannot set for PXE boot."
-                	continue
-        	fi
+		if [[ -z "$Kernel" || ! -f "$Kernel" || -z "$Initrd" || ! -f "$Initrd" ]]; then
+			echo "WARNING: Missing kernel or initrd file. Cannot set for PXE boot."
+			continue
+		fi
 
         	mkdir -p /tftpboot/${Moon_DeviceID}
         	ln -sf "$Kernel" /tftpboot/${Moon_DeviceID}/vmlinuz
         	ln -sf "$Initrd" /tftpboot/${Moon_DeviceID}/initrd.img
+
+		## these are needed because by default they see the kernel running on the core,
+		## which may be different from the one installed on the MD, thus not doing what we want by default
+		chroot /usr/pluto/diskless/"$DeviceID" depmod "$Kver"
+		chroot /usr/pluto/diskless/"$DeviceID" update-initramfs -k "$Kver" -u
 	else
 		for line in $Moon_DisklessImages; do
 			NameKernel="${line%%=*}"
